refactor namespace and packages, added fallback invoke into Launched block, renamed the Module
