--- plugins/AuthUI/build.gradle.kts
@@ -1,58 +1,0 @@
-plugins {
-    alias(libs.plugins.kotlin.multiplatform)
-    alias(libs.plugins.android.library)
-    alias(libs.plugins.compose)
-}
-
-description = "Extends gotrue-kt with ui composables"
-
-repositories {
-    mavenCentral()
-    maven("https://maven.pkg.jetbrains.space/public/p/compose/dev")
-}
-
-@OptIn(org.jetbrains.kotlin.gradle.ExperimentalKotlinGradlePluginApi::class)
-kotlin {
-    targetHierarchy.default()
-    jvmToolchain(8)
-    androidTarget {
-        publishLibraryVariants("release", "debug")
-    }
-
-    sourceSets {
-        all {
-            languageSettings.optIn("kotlin.RequiresOptIn")
-            languageSettings.optIn("io.github.jan.supabase.annotations.SupabaseInternal")
-            languageSettings.optIn("io.github.jan.supabase.annotations.SupabaseExperimental")
-        }
-        val commonMain by getting {
-            dependencies {
-                api(project(":"))
-                api(project(":gotrue-kt"))
-                implementation(compose.ui)
-                implementation(compose.runtime)
-            }
-        }
-        val androidMain by getting {
-            dependsOn(commonMain)
-            dependencies {
-                api(libs.android.play.store.auth)
-                implementation(libs.kotlinx.coroutines.play.services)
-                implementation(libs.androidx.activity.compose)
-            }
-        }
-    }
-}
-
-android {
-    namespace = "io.github.temk0.supabase.authui"
-    compileSdk = 33
-    sourceSets["main"].manifest.srcFile("src/androidMain/AndroidManifest.xml")
-    defaultConfig {
-        minSdk = 21
-    }
-    compileOptions {
-        sourceCompatibility = JavaVersion.VERSION_1_8
-        targetCompatibility = JavaVersion.VERSION_1_8
-    }
-}
