--- plugins/AuthUI/src/androidUnitTest/kotlin/io/github/temk0/supabase/authui/androidTest.kt
@@ -1,5 +1,0 @@
-package io.github.temk0.supabase.authui
-
-class AndroidGreetingTest {
-
-}
