--- plugins/AuthUI/src/commonMain/kotlin/io/github/temk0/supabase/authui/ui/GoogleButton.kt
@@ -1,7 +1,0 @@
-package io.github.temk0.supabase.authui.ui
-
-import androidx.compose.runtime.Composable
-import io.github.temk0.supabase.authui.AuthUI
-
-@Composable
-expect fun AuthUI.rememberLoginWithGoogle(onResult: (NativeSignInResult) -> Unit = {} , fallback: () -> Unit = {/*use gotrue.loginWith(Google) as default*/}): NativeSignInState
