--- plugins/AuthUI/src/commonTest/kotlin/io/github/temk0/supabase/authui/commonTest.kt
@@ -1,5 +1,0 @@
-package io.github.temk0.supabase.authui
-
-class CommonGreetingTest {
-
-}
