+++ plugins/ComposeAuth/src/commonMain/kotlin/io/github/jan/supabase/composeAuth/composable/GoogleLogin.kt
@@ -1,0 +1,7 @@
+package io.github.jan.supabase.composeAuth.composable
+
+import androidx.compose.runtime.Composable
+import io.github.jan.supabase.composeAuth.ComposeAuth
+
+@Composable
+expect fun ComposeAuth.rememberLoginWithGoogle(onResult: (NativeSignInResult) -> Unit = {}, fallback: () -> Unit = {}): NativeSignInState
